% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misinfo_exposure_score.R
\name{misinfo_exposure_score}
\alias{misinfo_exposure_score}
\title{Misinformation-Exposure Score Function
This function calculates the a minformation exposure score based on Twitter account that a user follows}
\arguments{
\item{user_id}{or screen_name of users for whom you want to calculate the scores}
}
\value{
users and their misinformation-exposure scores
}
\description{
Misinformation-Exposure Score Function
This function calculates the a minformation exposure score based on Twitter account that a user follows
}
